Task 4.4 Display favorite tasks in a separate section of the UI
Implementation of a tag filter to display only favourite tasks in the user interface.

The page is set up to automatically reload the user interface if you are in the favourite tag and uncheck a task.

In addition, the main tag for all tasks has been sorted to display favourite tasks first.