Optional task
Improved organisation of code folders for data, utilities, models, and routes. In addition, the list has been designed so that it does not exceed the height of the page and becomes a scrollable list.

Finally, to improve testing, I have made it possible to reload the backend while the code is running.